% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamedatautils.R
\name{add_team_score_label}
\alias{add_team_score_label}
\title{add a game label to team_score, good for x-axis.}
\usage{
add_team_score_label(team_score_subset, our_team = "NJD")
}
\arguments{
\item{team_score_subset}{Excerpt of team_score, filtered already for a team}

\item{our_team}{3-letter team_short format for default perspective}
}
\value{
game_info data frame with additional fields of our_ha, row_ha, col_ha, our_team, their_team
}
\description{
add a game label to team_score, good for x-axis.
}

