% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scutils.R
\name{aggregate_sc_h2h}
\alias{aggregate_sc_h2h}
\title{Aggregate sc_h2h}
\usage{
aggregate_sc_h2h(sc_h2h, rosters_C)
}
\arguments{
\item{sc_h2h}{Ev5on5 SC h2h data across multiple games}

\item{rosters_C}{Roster augmented with C info}
}
\value{
sc_h2h A summarized sc_h2h
}
\description{
Aggregate sc_h2h
}

