% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftintervalutils.R
\name{aggregate_toi_h2h}
\alias{aggregate_toi_h2h}
\title{Aggregate toi_h2h}
\usage{
aggregate_toi_h2h(toi_h2h_ev, rosters_C)
}
\arguments{
\item{toi_h2h_ev}{Ev5on5 toi_h2h data across multiple games}

\item{rosters_C}{Roster augmented with C info}
}
\value{
toi_h2h A summarized toi_h2h
}
\description{
Aggregate toi_h2h
}

