% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2hutils.R
\name{augment_rosters_C}
\alias{augment_rosters_C}
\title{Augment roster with faceoff count, center designation, num_last_name}
\usage{
augment_rosters_C(roster, pbp_df, player_tbl, center_fo_cutoff = 4)
}
\arguments{
\item{roster}{Data frame as from stage_game for arbitrary number of games}

\item{pbp_df}{Data frame of play by play}

\item{player_tbl}{Plyaer table}

\item{center_faceoff_rank_cutoff}{Cutoff for number of C per team}
}
\value{
Data frame of rosters with additional columns: faceoff count, is_center, num_last_name, L/R
}
\description{
Augment roster with faceoff count, center designation, num_last_name
}

