% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamedatautils.R
\name{calc_age}
\alias{calc_age}
\title{Calculate age given current date and birth date.
I can't believe I wrote a function for this.}
\usage{
calc_age(current_date, birth_date, digits = 0)
}
\arguments{
\item{current_date}{Date}

\item{birth_date}{Birth date}

\item{digits}{Number of rounding digits.  Default is 0.}
}
\value{
numeric
}
\description{
Calculate age given current date and birth date.
I can't believe I wrote a function for this.
}

