% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerstatsfn.R
\name{calc_shot_stats_basic}
\alias{calc_shot_stats_basic}
\title{calc_shot_stats_basic ---------------------------------------------------
Derived shot statistics for indivdidual.}
\usage{
calc_shot_stats_basic(game_player)
}
\arguments{
\item{game_player}{Data frame of player stats}
}
\value{
Data frame augment with more columns
}
\description{
Adapted from game_player_write_to_db_fn.R
}

