% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerstatsfn.R
\name{calc_shot_stats_rel}
\alias{calc_shot_stats_rel}
\title{Calculate relative shot stats}
\usage{
calc_shot_stats_rel(game_player, corsi_adj = 0.15, fenwick_adj = 0.1,
  shot_adj = 0.05)
}
\arguments{
\item{game_player}{Data frame}

\item{corsi_adj}{double.  Default = 0.15.}

\item{fenwick_adj}{double.  Default = 0.10.}

\item{shot_adj.}{double.  Default = 0.05.}
}
\value{
data frame
}
\description{
Calculate relative shot stats
}

