% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{create_heatmap_ggplot_from_h2h}
\alias{create_heatmap_ggplot_from_h2h}
\title{Create heatmap ggplot object from h2h data}
\usage{
create_heatmap_ggplot_from_h2h(h2h, value_type = "TOI", row_num_last_names,
  col_num_last_names, row_axis_title = "", col_axis_title = "",
  chart_title = "")
}
\arguments{
\item{h2h}{Data frame of num_last_name_1, num_last_name_2 pairs with value}

\item{value_type}{"TOI" or other}

\item{row_num_last_names}{Sorted vector of num_last_name. All will appear even if not in h2h}

\item{col_num_last_names}{Sorted vector of num_last_name  All will appear even if not in h2h}

\item{row_axis_title}{String}

\item{col_axis_title}{String}

\item{chart_title}{String}
}
\value{
ggplot object
}
\description{
Create heatmap ggplot object from h2h data
}

