% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{create_shotcolor_line_chart}
\alias{create_shotcolor_line_chart}
\title{Create Shot Color line chart.}
\usage{
create_shotcolor_line_chart(shots_df, goals_df, mandown_intervals_df, game_info,
  shift_interval_df, ev5on5 = FALSE, shotcolor = "ALL",
  include_blocked = F, fig_dir = NULL)
}
\arguments{
\item{shots_df}{Data frame from Green shot quality project}

\item{goals_df}{Data frame of goals.}

\item{mandown_intervals_df}{Data frame of when each team is down.}

\item{game_info}{Game info from stage_game.}

\item{shift_interval_df}{from stage_shift_interval}

\item{ev5on5}{Filters shots by ev5on5 or not. Default is FALSE.}

\item{shotcolor}{Chance type.  ALL, GREEN, RED, BLUE, WHITE.}

\item{include_blocked}{If \code{TRUE}, includes Blocked shots.}

\item{fig_dir}{}
}
\value{
None.  ggplot chart created.
}
\description{
Create Shot Color line chart.
}

