% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamedatautils.R
\name{get_goals_from_pbp}
\alias{get_goals_from_pbp}
\title{Get data frame of goals from playbyplay for a single game.
 Handles SO and OT goal cases.
 For SO goal, grabs arbitrary goal from winning team.}
\usage{
get_goals_from_pbp(pbp_df, game_info)
}
\arguments{
\item{pbp_df}{A data frame of playbyplay}

\item{game_info}{A data frame except of a single game from stage_game}
}
\value{
Data frame of goals with time as start_cum
}
\description{
Get data frame of goals from playbyplay for a single game.
 Handles SO and OT goal cases.
 For SO goal, grabs arbitrary goal from winning team.
}

