% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2hutils.R
\name{get_linemates_by_game}
\alias{get_linemates_by_game}
\title{Get linemates by game}
\usage{
get_linemates_by_game(this_player_name, this_season = "20152016",
  this_session_id = "2", player_tbl, game_h2h, this_player_team_games)
}
\arguments{
\item{this_player_name}{String of first_last_name (capitalize first letters only)}

\item{this_season}{Season string, "20152016" by default.}

\item{this_session_id}{Session ID, "2" by default.}

\item{player_tbl}{tbl of "player" from db.  already collected.}

\item{game_h2h}{tbl of "game_h2h".  Uncollected.}

\item{team_score}{tbl of "team_score".  Uncollected.}
}
\value{
data frame with game_number, num_last_name_2_label, toi_pct
}
\description{
Get linemates by game
}

