% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftintervalutils.R
\name{get_mandown_intervals}
\alias{get_mandown_intervals}
\title{Get intervals for when each team is shorthanded during a game.}
\usage{
get_mandown_intervals(shift_interval_df, game_info)
}
\arguments{
\item{shift_interval_df}{A data frame of shift intervals, from stage_shift_interval tbl}

\item{game_info}{A data frame of game info from stage_game}
}
\value{
Data frame of all man down time intervals from a game.
}
\description{
Get intervals for when each team is shorthanded during a game.
}

