% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamedatautils.R
\name{get_num_player_intervals_ha}
\alias{get_num_player_intervals_ha}
\title{Get all time intervals from a game with diff number of on ice players}
\usage{
get_num_player_intervals_ha(shift_interval_df, ha = "H")
}
\arguments{
\item{shift_interval_df}{A data frame of shift intervals, from stage_shift_interval tbl}

\item{ha}{A character "H" or "A"}
}
\value{
Data frame of time intervals
}
\description{
Get all time intervals from a game with diff number of on ice players
}

