% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamedatautils.R
\name{get_player_game_dates}
\alias{get_player_game_dates}
\title{Get player game dates for a single season.}
\usage{
get_player_game_dates(player_gp, team_score, recent_season_db)
}
\arguments{
\item{player_gp}{game_player tbl already filtered for a specific player of interest}

\item{team_score}{team_score tbl, not yet filtered for team of interest}

\item{recent_season_db}{Season string of form "20152016"}
}
\value{
list of df of game_dates and df of when player changed teams
}
\description{
Get player game dates for a single season.
}

