% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2hutils.R
\name{get_player_stats_by_game}
\alias{get_player_stats_by_game}
\title{Get player stats by game}
\usage{
get_player_stats_by_game(this_first_last_name, this_season = "20152016",
  this_session_id = "2", player_tbl, game_player, team_score,
  overwrite_ev5on5_toi = TRUE)
}
\arguments{
\item{this_first_last_name}{String of player first last name}

\item{this_season}{String like "20152016"}

\item{this_session_id}{"2" or "3"}

\item{player_tbl}{Table df}

\item{game_player}{game_player table, NOT collected}

\item{team_score}{team_score table, NOT collected}

\item{overwrite_ev5on5_toi}{Flag for whether to overwrite ev5on5 TOI with all - sh - pp TOI. Default is TRUE.}
}
\value{
data frame
}
\description{
Get player stats by game
}

