% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftintervalutils.R
\name{get_toi_matrix}
\alias{get_toi_matrix}
\title{Get TOI Matrix}
\usage{
get_toi_matrix(roster, shared_toi_ev)
}
\arguments{
\item{roster}{Roster for a game, as pulled from stage_game tbl}

\item{shared_toi_ev}{Shift interval, as pulled from stage_shift_interval, filtered for EV5on5.}
}
\value{
matrix of TOI with ha_number as row and col names
}
\description{
Get TOI Matrix
}

