% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scutils.R
\name{left_join_scf_i}
\alias{left_join_scf_i}
\title{left join scf_i onto ha_number scoring chance df.  Append a row for team level, H and A.}
\usage{
left_join_scf_i(on_ice_sc_by_ha_number, shots_df_filter)
}
\arguments{
\item{on_ice_sc_by_ha_number}{On ice scoring chance summary}

\item{shots_df_filter}{Filtered master shots df}
}
\value{
An expanded scoring chance df by ha_number, with scf_i added.
}
\description{
left join scf_i onto ha_number scoring chance df.  Append a row for team level, H and A.
}

