% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{move_heatmap_xaxis_to_top}
\alias{move_heatmap_xaxis_to_top}
\title{Move heatmap x-axis title and text to top of chart}
\usage{
move_heatmap_xaxis_to_top(p.mat)
}
\arguments{
\item{p.mat}{heatmap ggplot}
}
\value{
ggplot_gtable
}
\description{
Move heatmap x-axis title and text to top of chart
}

