% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/njdanalytics.R
\docType{package}
\name{njdanalytics-package}
\alias{njdanalytics}
\alias{njdanalytics-package}
\title{njdanalytics: A package for working with NHL.com data.}
\description{
The njdanalytics package provides functions for working in the
NJ Devils Hockey Analytics deparment
}
\details{
Details
}

