% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/njdanalytics.R
\name{njdanalytics}
\alias{njdanalytics}
\title{njdanalytics: A package for working with data scraped from NHL.com}
\description{
The njdanalytics package provides functions to work in the
NJ Devils Hockey Analytics deparment
}

