% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{rbind_plots}
\alias{rbind_plots}
\title{rbind ggplot charts}
\usage{
rbind_plots(plot_list, plot_heights)
}
\arguments{
\item{plot_list}{List of ggplot charts, top to bottom}

\item{plot_heights}{Vector of plot heights, top to bottom}
}
\value{
nothing
}
\description{
Bind mutliple ggplot charts vertically, aligning x axis.
Works best if x-axis is the same for all charts.
}

