% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/njdanalytics.R
\name{setup_nhl_db}
\alias{setup_nhl_db}
\title{Setup a src to our NHL postgreSQL database}
\usage{
setup_nhl_db()
}
\value{
An src for connecting to the 'nhl' PostgreSQL database.
}
\description{
Setup a src to our NHL postgreSQL database
}

