% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerstatsfn.R
\name{spread_team_comp_results}
\alias{spread_team_comp_results}
\title{spread_team_comp_results ----------------------------------------------------
Adapted from game_player_write_to_db_fn.R / spread_group_results()
 simplified here
      - don't have to deal with team perspective. already done for us
          - nhl_id instead of ha_number
 So while identical in practical way, needed a new version here}
\usage{
spread_team_comp_results(player_team_comp, prefix_base = "toi", fd_flag = T)
}
\arguments{
\item{player_team_comp}{Data frame of player level results to spread}

\item{prefix_base}{String for prefix for}

\item{fd_flag}{Flag for separating F from D}
}
\value{
Data frame with results spread out
}
\description{
spread_team_comp_results ----------------------------------------------------
Adapted from game_player_write_to_db_fn.R / spread_group_results()
 simplified here
      - don't have to deal with team perspective. already done for us
          - nhl_id instead of ha_number
 So while identical in practical way, needed a new version here
}

