% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamedatautils.R
\name{supplement_game_info}
\alias{supplement_game_info}
\title{Add our/their perspective fields to game_info}
\usage{
supplement_game_info(game_info, our_team = "NJD")
}
\arguments{
\item{game_info}{A single row from stage_game}

\item{our_team}{3-letter team_short format for default perspective}
}
\value{
game_info data frame with additional fields of our_ha, row_ha, col_ha, our_team, their_team
}
\description{
Add our/their perspective fields to game_info
}

