% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamedatautils.R
\name{tally_for_against_by_ha_number}
\alias{tally_for_against_by_ha_number}
\title{Count for and against stats by player (ha_number) for a
 single game.}
\usage{
tally_for_against_by_ha_number(pbp, output_col_names = c("f", "a", "net",
  "total"), reduce_cols = TRUE)
}
\arguments{
\item{pbp}{A data frame of playbyplay or shots}

\item{output_col_names}{A vector of names for output table}

\item{reduce_cols}{Return just results.  True by default}
}
\value{
data frame with columns ha_number, for_col, away_col
}
\description{
Count for and against stats by player (ha_number) for a
 single game.
}

