% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamedatautils.R
\name{tally_ha_number_for_against}
\alias{tally_ha_number_for_against}
\title{Count for and against stats by player (ha_number) for a
 single game.}
\usage{
tally_ha_number_for_against(pbp, output_col_names = c("for", "against"),
  reduce_cols = TRUE)
}
\arguments{
\item{pbp}{A data frame of playbyplay or shots}

\item{output_col_names}{A vector of names for output table}

\item{reduce_cols}{Return just results.  True by default}
}
\value{
data frame with columns ha_number, for_col, away_col
}
\description{
Count for and against stats by player (ha_number) for a
 single game.
}

