% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scutils.R
\name{tally_sc_by_ha_number}
\alias{tally_sc_by_ha_number}
\title{Tally scoring chances by ha_number}
\usage{
tally_sc_by_ha_number(shots_df_filter)
}
\arguments{
\item{shots_df_filter}{Shots df filtered for strength or shot color.}
}
\value{
data frame of ha_number, scf, sca, sc_net, sc_total, scf_i.
}
\description{
Tally scoring chances by ha_number
}

