% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamedatautils.R
\name{get_pairs_of_ha_numbers}
\alias{get_pairs_of_ha_numbers}
\title{Get every pair combination from a vector of players.}
\usage{
get_pairs_of_ha_numbers(ha_numbers, symmetric = TRUE)
}
\arguments{
\item{ha_numbers}{Vector of ha_number, already sorted alphabetically}

\item{symmetric}{Flag.  if TRUE, include (a,a), (a,b), (b,a), (b,b).  if FALSE, only (a,b).  TRUE by default.}
}
\value{
list of every pairwise combination
}
\description{
Get every pair combination from a vector of players.
}

