% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftintervalutils.R
\name{get_toi_matrix}
\alias{get_toi_matrix}
\title{Get EV5on5 TOI Matrix}
\usage{
get_toi_matrix(roster, shared_toi)
}
\arguments{
\item{roster}{Roster for a game including Goalies, as pulled from stage_game tbl}

\item{shared_toi}{Shift interval, as pulled from `stage_shift_interval`.
Usually filtered for EV5on5.}
}
\value{
matrix TOI with ha_number as row and col names
}
\description{
Get EV5on5 TOI Matrix
}

